- fix building on x86_64: "long" is 64bit there and according to the   comment, we want 32bit only. So we use "int"

+++ src/fpm.h
@@ -69,9 +69,9 @@ typedef unsigned long long ufpm;
 /*
  * The standard math should function with only 32 bits
  */
-typedef long fpm;
-typedef signed long sfpm;
-typedef unsigned long ufpm;
+typedef int fpm;
+typedef signed int sfpm;
+typedef unsigned int ufpm;
 #define FPM_BIT 10
  
 #endif
